third_party/WebKit/LayoutTests/fast/events/pointerevents/mouse-node-remove-expected.txt

Verifies the compatibility mouse events are sent correctly when the node is deleted on pointer event handler.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


 ======= Nothing is removed  =======
   **** mouse move *****
green received pointermove
green received mousemove
   **** mouse down *****
green received pointerdown
green received mousedown
   **** mouse up *****
green received pointerup
green received mouseup

 ======= Remove green on pointerdown  =======
   **** mouse move *****
green received pointermove
green received mousemove
   **** mouse down *****
green received pointerdown
     ==> Green box removed
grey received mousedown
   **** mouse up *****
grey received pointerup
grey received mouseup

 ======= Remove green on pointerup  =======
   **** mouse move *****
green received pointermove
green received mousemove
   **** mouse down *****
green received pointerdown
green received mousedown
   **** mouse up *****
green received pointerup
     ==> Green box removed
grey received mouseup

 ======= Remove green on pointermove  =======
   **** mouse move *****
green received pointermove
     ==> Green box removed
grey received mousemove
   **** mouse down *****
grey received pointerdown
grey received mousedown
   **** mouse up *****
grey received pointerup
grey received mouseup

PASS successfullyParsed is true

TEST COMPLETE
